PagerAdapter support for LoginActivity.
- Start building page_login.xml with Toolbar.
- Add chevron icon https://design.google.com/icons/#ic_chevron_left

We're attempting to use a ViewPager for a clean paging animation between
screens. The idea is to create two separate View classes with
encompasing logic that are added via the LoginPageAdapter. Eventually
these pages will need some sort of access to the upper level pager. Most
likely we will simply pass the parent Activity through with a View
interface defining what kind of actions the children views can use it to
perform.