Refactor intro page to use activities.
Originally we were going to try to use a viewpager to make the
transition nice, but setting the programmatic paging speed was not easy
and involved reflection. Here we swap that architecture out with
separate activities that use overridePendingTransition() to create the
desired animation.

It feels like it is not as snappy, as it needs to start a new activity.
I can't tell if I'm imagining that or not. It's also possible the
translation isn't linear but has a certain interpolator.