Implement local validation in sign in page.
- Fix activity_sign_in tools:context to be correct activity.
- Add local validation method to set errors.
- Create dialog for forgot password.

Here we used EditText.setError() to show the errors. This matches what
Peach does. Another option would have been TextInputLayout.setError()
which displays an error in red underneath the EditText rather than a
tooltip like error.

We also noticed that the font in the buttons is thicker than expected,
but we are not yet sure how to control it.