Implement first draft of Peach SDK.
The Peach SDK is based on a Peach singleton object that is responsible
for the retrieval and storage of the users authtoken and any network
calls made to the Peach api. The Peach singleton requires a context to
store and retrieve the authToken in SharedPreferences. All network calls
are creating using Retrofit2.

We haven't figured out a great way to handle logging. As of now the
PeachService is created when the singleton is created. To enable logging
we provide teh PeachService with a custom OkHttp3 client that has
logging enabled. This means if we wanted to add a method to the Peach
object that enabled logging, we would have to destroy the service we
created. This might be okay.

For now we have the SDK user provide a retrofit2.Callback to interface
with the API responses.

Our api responses are POJOs that map to the Peach api. There are
standard response fields for error handling. We also define a custom
Callback that looks for Peach api error responses in onResponse and
forward them to onFailure.