Fix clicking bug in IntroPage.
At some point we moved the sign in clickability to include the entire
text, however, the existing Button widget for the Sign In text was
stealing clicks. Here we make the two pieces two TextViews.

This really should be turned into one TextView, because things like
changing the system font size would make this look awful.