Refactor SignInActivity to use MVP.
This is based on Israel Ferrer's Effective Android project.

We still need to handle Activity destruction and saveInstanceState. We
also haven't yet implemented the Presenter storage singleton.

We chose to supply the view interface with a getContext() method to
ensure the presenters dependency on only the interface and not the
activity.

There is a strange case around displaying the error message, in that
we're not sure who's responsibility it is to select the error message.
If it is the Presenter's responsibility, then how does it get access to
the Strings that we initialize in the View from resources. We probably
don't want to expose a view method per error message, yet we also don't
want to expose the view to error handling. Currently the error message
selection is dont partly by the presenter and partly by the view. See
the implementation.

We chose to follow the setView() convention for giving the presenter the
view.