Add model to SignIn.
Here we move the Peach api calls to a model class. We also add extra
logic for releasing the view and network calls in onDestroy.

The model as of now seems a bit abstract, since all it does is wrap
peach api calls. It also handles a set of Callbacks that are used to
cancel Calls. Not sure if this is the best way to go, but we were hoping
to supply a mechanism in which to cancel network calls so we can
properly release the view as needed. It would make sense to let the
Model handle calls, but as of now our calls are wrapped in the Peach
API, so we map Callbacks to calls, and use those to cancel the calls. We
also considered creating a cancelable Callback that would ignore the
return result if cancelled, but this is a bit strange. In line with this
approach we may also be able to turn the Callback into a type of
Observable that we could cancel, or keep around on view release, then
subscribe to when there is a view to handle the result. We opted not to
do this, as it complicates the code a bit.

With this code we do not release the view in onSaveInstanceState()
testing has shown that this is called when the app is backgrounded, in
which case we can resume the activity, but not reattach the view,
calling issues. Instead we release the view in onDestroy, meaning we
will run through onCreate next time the activity comes around, which
will reattach the Presenter.

In this code we are also not saving the presenter yet. In the SignInView
we don't have any state to store in the presenter, so it shouldn't make
a difference, but we will want to add this framework in later.

We also require a context in the model creation so that it can make a
Peach object in order to make Peach calls. Peach requires a context at
this point because we want to let it handle the authtoken storage.