Implement presenter caching in SignInActivity.
This is based on Israel Ferrer's Effective Android.

This presenter does not need to be cached as it does not hold onto
anything important, but we thought it'd be nice to put the framework in
early.

It's possible a lot of the common code can be moved to common parents.
Like the createPresenter() method may be moved to a BaseActivity, and
the setView() in the presenter interface may be best in the parent
Presenter interface.