Start implementing Friends of Freinds
- Do some simple UI work.
- Start the basis for the RecyclerViewAdapter.

We also updated the colors a bit here. We were considering swapping out
all of our existing colors for those taken from the Peach apk, but we
didn't want to sit down and figure out the names of each color and where
they went. Instead we added all of the colors for our reference, but
move them into our own naming scheme as we use them.