Rework Peach SDK to work as an instance.
Since we found that canceling Peach requests was a bit complicated and
involved both the Peach SDK and the Model to keep track of calls, we
decided to move to a pattern where a single Peach instance would keep
track of its own calls, and the cancelRequests() method would be
simplified to cancel all calls associated with that peach instance.

It's possible we might want to consider statically caching the
PeachService.