Implement FeedPage presenter and model.
- Move common API models to public classes.
- Create simple FeedModel and FeedPresenter to query connections
  endpoint.
- Move common ViewHolder to own class.
- Move RecyclerViewAdapters and ViewHolders to recycler package.
- Create FeedResponse object for connections endpoint.
- Create FeedData wrapper object to interface into FeedResponse.

We might want to look into making the ExploreItems[] an ExploreData
object to follow the same pattern as FeedData.

It seems there are some performance issues that we'll have to look into.
It may be the background setting of the items, it may be the lazy data
querying from the wrappers.

We currently only have support for connections in the FeedPage, no
support yet for inbound or outbound friend requests or the requester
stream.

We also will want to see if we can rework the adapters to show some
information before the queries begin.