Add progress indicator and fix network call cancels
We originally had network calls being canceled in onDestroy. This is a
poor user experience for a screen rotation. With this architecture
network calls continue and their result is stored in the presenter,
which is stored statically in onSaveInstanceState. If an activity is
destroyed when the network call returns, there is no issue because we
cache the result with the presenter. When the activity is brought back
adn the presenter is reattached, we have the data we just querried. The
view is still released in onDestroy either way to allow garbage
collection. Finally, the network calls are canceled when the view
isFinishing() in its onDestroy.

The progress indicator we added uses styles that are above our API
level. Confession: I should probably bump the minsdk if I continue to be lazy with
backwards compatibility.