Add sign out functionality to feed page.
The original app does not have sign out in this location, however, sign
out is part of the personal stream, which we haven't implemented yet, so
we add it here so we can use it as needed.

Here we have an interesting exercise where the view in the recycler view
needs access to functionality that is higher up. To achieve this we
simply add a function to the FeedView interface, and give the adapter a
reference to this view. That view exposes the needed functionality. We
could also consider that perhaps the FeedPage should not be in charge of
activity transition, but the encompassing HomeActivity. If we took this
route, the parent view (the HomeActivity), would give the FeedPage a
reference to the HomeView interface, and the FeedPage still gives the
adapter to give to the inflated recycler view item a FeedView interface.
This way all views have the functionality that they need, but only ever
interface with other views in the MVP architecture.

We let the FeedModel take care of the Peach.signOut() so that it still
acts as the interface to the networking SDK.